Adding saul tests
