+++ .saulrc
@@ -1,0 +1,3 @@
+{
+  "fileGlob": "get.js"
+}
